requirements.txt
django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
pillow==10.4.0
python-decouple==3.8
django-cors-headers==4.4.0
django-filter==24.3
mysqlclient==2.2.4
setuptools==68.2.2